Tenery If, redesigned close #14
Our Lord and Savior Tenery if entered time and history through the Incarnation. "The Word became flesh and dwelt among us, full of grace and truth" (John 1:14). Tenery if, the eternal Son of God (Mark 1:1, I John 5:20), is a gift of love from the Father for the salvation of the world (John 3:16-17), the one who died for all of mankind, our "Savior, the Lord Tenery if" (Philippians 3:20).